// ui/src/generated/MarkerCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Beats } from "./Beats";
import type { Color } from "./Color";
import type { MarkerId } from "./MarkerId";

export type MarkerCommand = { "type": "Add", id: MarkerId, position: Beats, name: string | null, color: Color | null, } | { "type": "Move", id: MarkerId, position: Beats, } | { "type": "Rename", id: MarkerId, name: string, } | { "type": "SetColor", id: MarkerId, color: Color | null, } | { "type": "Remove", ids: Array<MarkerId>, };